Fix html metadata tag encoding priority in HtmlEncodingDetector.java and revert the pull request #8.
