+++ tika-parent/pom.xml
@@ -33,7 +33,7 @@
 
   <groupId>com.github.lafa.tikaNoExternal</groupId>
   <artifactId>tika-parent</artifactId>
-  <version>1.0.7</version>
+  <version>1.0.8</version>
   <packaging>pom</packaging>
 
   <name>Apache Tika parent</name>
