+++ tika-translate/pom.xml
@@ -25,7 +25,7 @@
   <parent>
     <groupId>com.github.lafa.tikaNoExternal</groupId>
     <artifactId>tika-parent</artifactId>
-    <version>1.0.7</version>
+    <version>1.0.8</version>
     <relativePath>../tika-parent/pom.xml</relativePath>
   </parent>
